+++ OAuth/ResourceOwner/FacebookResourceOwner.php
@@ -37,7 +37,16 @@ class FacebookResourceOwner extends GenericOAuth2ResourceOwner
      */
     public function getAuthorizationUrl($redirectUri, array $extraParameters = array())
     {
-        return parent::getAuthorizationUrl($redirectUri, array_merge(array('display' => $this->options['display']), $extraParameters));
+        $extraOptions = array();
+        if (isset($this->options['display'])) {
+            $extraOptions['display'] = $this->options['display'];
+        }
+
+        if (isset($this->options['auth_type'])) {
+            $extraOptions['auth_type'] = $this->options['auth_type'];
+        }
+
+        return parent::getAuthorizationUrl($redirectUri, array_merge($extraOptions, $extraParameters));
     }
 
     /**
@@ -89,11 +98,12 @@ class FacebookResourceOwner extends GenericOAuth2ResourceOwner
             'use_commas_in_scope' => true,
 
             'display'             => null,
+            'auth_type'           => null,
         ));
 
         $resolver->setAllowedValues(array(
-            // @link https://developers.facebook.com/docs/reference/dialogs/#display
-            'display' => array('page', 'popup', 'touch', null),
+            'display' => array('page', 'popup', 'touch', null), // @link https://developers.facebook.com/docs/reference/dialogs/#display
+            'auth_type' => array('rerequest', null), // @link https://developers.facebook.com/docs/reference/javascript/FB.login/
         ));
     }
 }