+++ OAuth/ResourceOwner/GitHubResourceOwner.php
@@ -60,8 +60,36 @@ class GitHubResourceOwner extends GenericOAuth2ResourceOwner
             'access_token_url'    => 'https://github.com/login/oauth/access_token',
             'revoke_token_url'    => 'https://api.github.com/applications/%s/tokens/%s',
             'infos_url'           => 'https://api.github.com/user',
+            'emails_url'          => 'https://api.github.com/user/emails',
 
             'use_commas_in_scope' => true,
         ));
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getUserInformation(array $accessToken, array $extraParameters = array())
+    {
+        $response = parent::getUserInformation($accessToken, $extraParameters);
+
+        $responseData = $response->getResponse();
+
+        if (empty($responseData['email'])) {
+            // fetch the email addresses linked to the account
+            $content = $this->httpRequest($this->normalizeUrl($this->options['emails_url']), null, array('Authorization: Bearer '.$accessToken['access_token']));
+
+            foreach ($this->getResponseContent($content) as $email) {
+                if (!empty($email['primary'])) {
+                    // we only need the primary email address
+                    $responseData['email'] = $email['email'];
+                    break;
+                }
+            }
+
+            $response->setResponse($responseData);
+        }
+
+        return $response;
     }
 }