+++ Tests/OAuth/ResourceOwner/FacebookResourceOwnerTest.php
@@ -40,6 +40,26 @@ json;
         $this->resourceOwner->getAccessToken($request, 'http://redirect.to/');
     }
 
+    public function testAuthTypeRerequest()
+    {
+        $resourceOwner = $this->createResourceOwner($this->resourceOwnerName, array('auth_type' => 'rerequest'));
+
+        $this->assertEquals(
+            $this->options['authorization_url'] . '&response_type=code&client_id=clientid&state=random&redirect_uri=http%3A%2F%2Fredirect.to%2F&auth_type=rerequest',
+            $resourceOwner->getAuthorizationUrl('http://redirect.to/')
+        );
+    }
+
+    public function testAuthTypeRerequestAndDisplayPopup()
+    {
+        $resourceOwner = $this->createResourceOwner($this->resourceOwnerName, array('display' => 'popup', 'auth_type' => 'rerequest'));
+
+        $this->assertEquals(
+            $this->options['authorization_url'] . '&response_type=code&client_id=clientid&state=random&redirect_uri=http%3A%2F%2Fredirect.to%2F&display=popup&auth_type=rerequest',
+            $resourceOwner->getAuthorizationUrl('http://redirect.to/')
+        );
+    }
+
     public function testDisplayPopup()
     {
         $resourceOwner = $this->createResourceOwner($this->resourceOwnerName, array('display' => 'popup'));